requirements.txt
selenium~=4.26.1
pandas~=2.2.3
webdriver-manager~=4.0.2
Pillow~=11.0.0

SpeechRecognition~=3.11.0
requests~=2.32.3
pydub~=0.25.1